Split code to more files, no changes
