+++ src/environment.h
@@ -1,0 +1,6 @@
+#pragma once
+
+typedef struct {
+        sd_bus *bus;
+        const char *dbus_prefix;
+} Environment;
